refactor(header): dataNavArray move to utils
