+++ src/utils/dataNavArray.js
@@ -1,0 +1,22 @@
+const dataNavArray = [
+	{
+		id: 0,
+		label: 'проекты',
+		path: '/projects',
+		anchor: '',
+	},
+	{
+		id: 1,
+		label: 'новости',
+		path: '/',
+		anchor: 'news',
+	},
+	{
+		id: 2,
+		label: 'связаться с нами',
+		path: '/',
+		anchor: 'request',
+	},
+];
+
+export default dataNavArray;
